+++ packages/api/CHANGELOG.md
@@ -1,5 +1,13 @@
 # ngn-api
 
+## 0.1.10
+
+### Patch Changes
+
+- Plugin system
+- Updated dependencies
+  - ngn-persistence@0.1.10
+
 ## 0.1.9
 
 ### Patch Changes