+++ packages/os/CHANGELOG.md
@@ -1,5 +1,11 @@
 # ngn-os
 
+## 0.1.10
+
+### Patch Changes
+
+- Plugin system
+
 ## 0.1.9
 
 ### Patch Changes