+++ packages/persistence/CHANGELOG.md
@@ -1,5 +1,13 @@
 # ngn-persistence
 
+## 0.1.10
+
+### Patch Changes
+
+- Plugin system
+- Updated dependencies
+  - ngn-os@0.1.10
+
 ## 0.1.9
 
 ### Patch Changes